Starting to move tools into RDGInspection.h
